class SAFHEH_AVAL
Log in for employees.
Add new user.
Saved users in the file.
Hashed password (optional)
Display the inventory of goods.
Issuing an invoice to the customer.
Each invoice is saved as a csv file (research about its structure)
The goods that are issued in the invoice must be subtracted from the inventory.
Saved history of invoices in the file.
Stored inventory of goods in the file.
The ability to use different currencies such as euros and dollars and convert them.
Ability to update exchange rates manually.
Update the exchange rate automatically API (optional)